Restructure rules files
